BUG: MANTIS-908 (persistence of gamma); ENH: Updated MVD2 version to 0.7.3; updated MVD2 data version to 0.7.3; REFAC: Refactored ColorDynamicsController/Widget gamma functions (factorized log/exp formula into Widget, added missing disconnect gamma-slots); Removed debug traces.

+++ Code/Common/Gui/mvdColorDynamicsController.h
@@ -251,7 +251,7 @@ private:
   /**
    */
   void SetNoData();
-  
+
   /**
    */
   void SetGamma();
@@ -357,9 +357,9 @@ private slots:
   /**
    * \brief Slot called when gamma value is changed
    *
-   * \param value Gamma value
+   * \param value Gamma value.
    */
-  void OnGammaValueChanged(int value);
+  void OnGammaValueChanged( double value );
 
 };
